// ms-users/src/domain/exceptions/enums/authentication-exception.enum.ts
export enum AuthenticationExceptionEnum {
  USER_CREDENTIAL_IS_NOT_PROTECTED = 'USER_CREDENTIAL_IS_NOT_PROTECTED',
  VERIFICATION_FAILED = 'VERIFICATION_FAILED',
  TOKEN_SUBSTITUTION = 'TOKEN_SUBSTITUTION',
  HACKING_ATTEMPT_DETECTED = 'HACKING_ATTEMPT_DETECTED',
  THIRD_PARTY_CLIENT = 'THIRD_PARTY_CLIENT',
  THIRD_PARTY_IP = 'THIRD_PARTY_IP',
  EMAIL_IS_NOT_FOUND = 'EMAIL_IS_NOT_FOUND',
  EMAIL_IS_NOT_FOUND_IN_LOGIN = 'EMAIL_IS_NOT_FOUND_IN_LOGIN',
  PHONE_IS_NOT_FOUND = 'PHONE_IS_NOT_FOUND',
  PHONE_IS_NOT_FOUND_IN_LOGIN = 'PHONE_IS_NOT_FOUND_IN_LOGIN',
  CONFIRMATION_TYPE_IS_NOT_FOUND = 'CONFIRMATION_TYPE_IS_NOT_FOUND',
  LIMIT_FOR_REQUEST_EXCEEDED = 'LIMIT_FOR_REQUEST_EXCEEDED',
  INVALID_ROLE_FOR_THIS_ACTION = 'INVALID_ROLE_FOR_THIS_ACTION',
  INCORRECT_PASSWORD = 'INCORRECT_PASSWORD',
  INCORRECT_OLD_PASSWORD = 'INCORRECT_OLD_PASSWORD',
  USER_DATA_UPDATE_FAILED_WITH_ERROR = 'USER_DATA_UPDATE_FAILED_WITH_ERROR',
  SESSION_IS_CURRENT = 'SESSION_IS_CURRENT',

  // JWT
  ACCESS_TOKEN_EXPIRED = 'ACCESS_TOKEN_EXPIRED',
  REFRESH_TOKEN_EXPIRED = 'REFRESH_TOKEN_EXPIRED',
  ACCESS_TOKEN_UNCORRECTED = 'ACCESS_TOKEN_UNCORRECTED',
  REFRESH_TOKEN_UNCORRECTED = 'REFRESH_TOKEN_UNCORRECTED',

  SESSION_NOT_FOUND = 'SESSION_NOT_FOUND',
  EMAIL_CODE_NOT_FOUND = 'EMAIL_CODE_NOT_FOUND',
  CODE_NOT_FOUND = 'CODE_NOT_FOUND',
  CODE_IS_NOT_ACCEPTED = 'CODE_IS_NOT_ACCEPTED',
  INCORRECT_CODE = 'INCORRECT_CODE',
  REGISTRATION_DATA_NOT_FOUND = 'REGISTRATION_DATA_NOT_FOUND',
  INCORRECT_REGISTRATION_CODE = 'INCORRECT_REGISTRATION_CODE',
  TOKEN_IS_MISSING_IN_HEADERS = 'TOKEN_IS_MISSING_IN_HEADERS',
}